Migrate script to luigi workflow
